requirements.txt
djangorestframework==3.12.4
dj-rest-auth